Функция создания мокового объекта перенесена из класса компонента Film в отдельный файл data.js
